% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycats.R
\name{mycats}
\alias{mycats}
\title{My Cats Function}
\usage{
mycats(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
mycats()
}
\keyword{cats}
